move helper functions to their own file
